==> 07-conclusion.Rmd <==
# Conclusion

By analyzing the dataset through visualization, we had a deeper understanding of the user behavior from the Taobao dataset.

Firstly, we observed that most of the user behavior consists of clicking an item. This is understandable, however, we surprisingly observe that buying consist of only 2% of the user behavior. This indicates the fundamental difference between online shopping and on-site shopping, where in the former case consumers can quickly browse various products and only choose to buy a small portion of them while in the latter case buying occurs more often. We further investigate the total number of user behaviors by hours, days and item category. We found that clicking and uying does not happen concurrently. People tend to view their products at night and make purchases in the morning or afternoon. Moreover, we also observed thatthe total clicking behavior occurs more frequently during the weekends while less frequently during the weekdays. However, for buying behaviors, it is the complete opposite. By this 2 observations, we can obtain a rough idea about the online shoppers' behavior. They prefer browing the products during weekend and nights, while buying occurs during weekdays and afternoon. Moreover, we analyze the user behavior by item category. Although the dataset provides no information about the specific item categories, we found one interesting pattern about the categories: the items that were viewed most frequently are the ones whose conversion rate is the lowest.

Then we further analyzed which action between listing and carting is more likely to indicate a user will finally buy the product. We found that carting a product significantly increases the probability that a user will finally buy a product, while listing a product only slightly increases the users' will to buy it. This also confirms with our personal experience, where we are more likely to cart the products we really like to buy, while keeping the products that we aren't ready to spend money on to the list.

Finally, we investigate the outliers of users. We define the outliers to be the ones with excessively active users. And we foudn that in fact many users exhibit abnormal behaviors with significantly more behaviors than the rest. We suspect that they are click-farming, which has always been a problem not only for Taobao but also other online shopping platform such as Amazon and Ebay. Unlukily, we are unable to draw the conclusion that those users are click-farming since they do not have a large volumn of buying activies. However, we do find interesting behaviors about those users: they are almost fanatic in viewing various products while being very unlikely to buy them. We observe 4 times more  clicking behaviors for the excessively active users but only 2 times more buying behaviors. We conclude that those users are simply addictive to clicking on different products and cosnider browing products a hobby for them.
